CART-89 build: Separate gurt/types.h from gurt/common.h
It turns out common.h includes debug which makes it tricky in
DAOS to include it and define its own DD_FAC.  This ensures
we don't include debug stuff unless we explicity include it.
